Move readCSV bench in converter_benchmark (#6)
