+++ cpp/src/arrow/util/CMakeLists.txt
@@ -64,7 +64,6 @@ add_arrow_benchmark(int_util_benchmark)
 add_arrow_benchmark(machine_benchmark)
 add_arrow_benchmark(number_parsing_benchmark)
 add_arrow_benchmark(range_benchmark)
-add_arrow_benchmark(read_csv_benchmark)
 add_arrow_benchmark(thread_pool_benchmark)
 add_arrow_benchmark(trie_benchmark)
 add_arrow_benchmark(utf8_util_benchmark)
